+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- Migrator service inside `php-web-server` repository.
+
 ### Removed
 
 - Consul script from `jenskins-slave` image.